chore(tests): move tests around
